requirements.txt
Django==1.11.2
mysqlclient==1.3.12
wheel==0.30.0
gunicorn==19.7.1
psycopg2==2.7.4
social-auth-app-django==2.1.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
django-tastypie==0.14.1